refactor(telegram): extract DM-topic machinery into TelegramDmTopicMixin (adapter god-file slice A1)
(cherry picked from commit a8890d948621159e505ac057ba7f2b7653ce258f)

# Conflicts:
#	plugins/platforms/telegram/adapter.py
