+++ plugins/platforms/telegram/adapter.py
@@ -302,6 +302,7 @@ from plugins.platforms.telegram.telegram_messaging import (
 from plugins.platforms.telegram.telegram_rich import TelegramRichMixin
 from plugins.platforms.telegram.telegram_polling import TelegramPollingMixin
 from plugins.platforms.telegram.telegram_lifecycle import TelegramLifecycleMixin
+from plugins.platforms.telegram.telegram_dm_topics import TelegramDmTopicMixin
 from utils import atomic_replace, env_float, env_int
 
 from plugins.platforms.telegram.telegram_inbound import (
@@ -479,7 +480,7 @@ class _PollingLifecycleAbort(RuntimeError):
     """Internal control flow for polling startup fenced by teardown."""
 
 
-class TelegramAdapter(TelegramMediaMixin, TelegramLifecycleMixin, TelegramReactionsMixin, TelegramPollingMixin, TelegramIngestMixin, TelegramTextDeliveryMixin, TelegramRichMixin, BasePlatformAdapter):
+class TelegramAdapter(TelegramMediaMixin, TelegramLifecycleMixin, TelegramReactionsMixin, TelegramPollingMixin, TelegramIngestMixin, TelegramTextDeliveryMixin, TelegramRichMixin, TelegramDmTopicMixin, BasePlatformAdapter):
     """
     Telegram bot adapter.
 
@@ -978,264 +979,6 @@ class TelegramAdapter(TelegramMediaMixin, TelegramLifecycleMixin, TelegramReacti
         # Unauthorized DM that the gateway would pair: forward so pairing can run.
         return self._should_pass_unauthorized_dm_for_pairing(source)
 
-    @classmethod
-    def _metadata_thread_id(cls, metadata: Optional[Dict[str, Any]]) -> Optional[str]:
-        if not metadata:
-            return None
-        thread_id = metadata.get("thread_id") or metadata.get("message_thread_id")
-        return str(thread_id) if thread_id is not None else None
-
-    @classmethod
-    def _metadata_direct_messages_topic_id(cls, metadata: Optional[Dict[str, Any]]) -> Optional[str]:
-        if not metadata:
-            return None
-        topic_id = metadata.get("direct_messages_topic_id") or metadata.get("telegram_direct_messages_topic_id")
-        return str(topic_id) if topic_id is not None else None
-
-    @classmethod
-    def _metadata_reply_to_message_id(cls, metadata: Optional[Dict[str, Any]]) -> Optional[int]:
-        if not metadata:
-            return None
-        reply_to = metadata.get("telegram_reply_to_message_id")
-        return int(reply_to) if reply_to is not None else None
-
-    @classmethod
-    def _is_private_dm_topic_send(
-        cls,
-        chat_id: str,
-        thread_id: Optional[str],
-        metadata: Optional[Dict[str, Any]],
-    ) -> bool:
-        if cls._metadata_direct_messages_topic_id(metadata) is not None:
-            return bool(
-                metadata
-                and metadata.get("telegram_dm_topic_reply_fallback")
-                and cls._metadata_reply_to_message_id(metadata) is not None
-            )
-        if metadata and metadata.get("telegram_dm_topic_created_for_send"):
-            return False
-        return bool(
-            thread_id
-            and metadata
-            and metadata.get("telegram_dm_topic_reply_fallback")
-        )
-
-    @staticmethod
-    def _dm_topic_missing_anchor_error() -> str:
-        return "Telegram DM topic delivery requires a reply anchor; refusing to send outside the requested topic"
-
-    @classmethod
-    def _reply_to_message_id_for_send(
-        cls,
-        reply_to: Optional[str],
-        metadata: Optional[Dict[str, Any]] = None,
-        reply_to_mode: Optional[str] = None,
-    ) -> Optional[int]:
-        if reply_to:
-            return int(reply_to)
-        if metadata and metadata.get("telegram_dm_topic_reply_fallback"):
-            if reply_to_mode == "off":
-                return None
-            return cls._metadata_reply_to_message_id(metadata)
-        return None
-
-    @classmethod
-    def _thread_kwargs_for_send(
-        cls,
-        chat_id: str,
-        thread_id: Optional[str],
-        metadata: Optional[Dict[str, Any]] = None,
-        reply_to_message_id: Optional[int] = None,
-        reply_to_mode: Optional[str] = None,
-    ) -> Dict[str, Any]:
-        """Return Telegram send kwargs for forum and direct-message topic routing.
-
-        Supergroup/forum topics use ``message_thread_id``. True Bot API Direct
-        Messages topics can opt in with explicit ``direct_messages_topic_id``
-        metadata. Hermes-created private-chat topic lanes are marked with
-        ``telegram_dm_topic_reply_fallback``. Live replies send the private
-        topic thread id together with a reply anchor; synthetic/resumed sends
-        without an anchor use ``direct_messages_topic_id`` when metadata has it.
-        ``message_thread_id`` alone can render outside the visible lane.
-
-        When ``reply_to_mode`` is ``"off"``, the reply anchor is suppressed for
-        DM topic fallback sends while preserving the ``message_thread_id`` so
-        the message still lands in the correct topic.
-        """
-        if metadata and metadata.get("telegram_dm_topic_reply_fallback"):
-            if reply_to_mode == "off":
-                return {"message_thread_id": cls._message_thread_id_for_send(thread_id)}
-            if reply_to_message_id is None:
-                reply_to_message_id = cls._metadata_reply_to_message_id(metadata)
-            if reply_to_message_id is None:
-                direct_topic_id = cls._metadata_direct_messages_topic_id(metadata)
-                if direct_topic_id is not None:
-                    return {
-                        "message_thread_id": None,
-                        "direct_messages_topic_id": int(direct_topic_id),
-                    }
-                return {}
-            return {"message_thread_id": cls._message_thread_id_for_send(thread_id)}
-        direct_topic_id = cls._metadata_direct_messages_topic_id(metadata)
-        if direct_topic_id is not None:
-            return {
-                "message_thread_id": None,
-                "direct_messages_topic_id": int(direct_topic_id),
-            }
-        return {"message_thread_id": cls._message_thread_id_for_send(thread_id)}
-
-    @classmethod
-    def _message_thread_id_for_send(cls, thread_id: Optional[str]) -> Optional[int]:
-        if not thread_id or str(thread_id) == cls._GENERAL_TOPIC_THREAD_ID:
-            return None
-        return int(thread_id)
-
-    @classmethod
-    def _message_thread_id_for_typing(cls, thread_id: Optional[str]) -> Optional[int]:
-        # Asymmetric with _message_thread_id_for_send on purpose. Telegram's
-        # sendMessage and sendChatAction treat thread id "1" (the forum General
-        # topic) differently: sends reject message_thread_id=1 and must omit it,
-        # but sendChatAction needs message_thread_id=1 to place the typing
-        # bubble in the General topic (omitting it hides the bubble entirely
-        # from the client's view of that topic). Preserve the real id here —
-        # sends still map "1" → None via _message_thread_id_for_send.
-        if not thread_id:
-            return None
-        return int(thread_id)
-
-    @staticmethod
-    def _is_thread_not_found_error(error: Exception) -> bool:
-        return "thread not found" in str(error).lower()
-
-    def _prune_stale_dm_topic_binding(
-        self, chat_id: Any, thread_id: Any,
-    ) -> None:
-        """Drop the stale ``telegram_dm_topic_bindings`` row for a
-        topic Telegram has confirmed deleted.
-
-        Without this prune the recovery logic in
-        ``gateway.run._recover_telegram_topic_thread_id`` keeps
-        steering future inbound messages to the dead thread (the
-        bug behind #31501 — tool progress, approvals, replies all
-        end up in the wrong place even though the user has moved
-        on to a fresh topic).  Best-effort: we never raise from a
-        send-fallback path — a failed cleanup must not turn into a
-        failed user-facing send.
-        """
-        if chat_id is None or thread_id is None:
-            return
-        store = getattr(self, "_session_store", None)
-        if store is None:
-            return
-        db = getattr(store, "_db", None)
-        if db is None or not hasattr(db, "delete_telegram_topic_binding"):
-            return
-        try:
-            removed = db.delete_telegram_topic_binding(
-                chat_id=str(chat_id), thread_id=str(thread_id),
-            )
-        except Exception:
-            logger.debug(
-                "[%s] delete_telegram_topic_binding failed for "
-                "chat=%s thread=%s — skipping prune",
-                self.name, chat_id, thread_id, exc_info=True,
-            )
-            return
-        if removed:
-            logger.info(
-                "[%s] Pruned stale Telegram DM topic binding "
-                "chat=%s thread=%s (Bot API: thread not found)",
-                self.name, chat_id, thread_id,
-            )
-
-    @staticmethod
-    def _is_bad_request_error(error: Exception) -> bool:
-        name = error.__class__.__name__.lower()
-        if name == "badrequest" or name.endswith("badrequest"):
-            return True
-        try:
-            from telegram.error import BadRequest
-            return isinstance(error, BadRequest)
-        except ImportError:
-            return False
-
-    @classmethod
-    def _should_retry_without_dm_topic_reply_anchor(
-        cls,
-        error: Exception,
-        metadata: Optional[Dict[str, Any]],
-        reply_to_message_id: Optional[int],
-    ) -> bool:
-        """True when a DM-topic send should be retried with routing stripped.
-
-        Two cases trigger the retry:
-
-        1. The original anchor-stale case — the reply target was deleted, so
-           Bot API returns "message to be replied not found". The retry drops
-           the reply anchor and the topic id together.
-
-        2. The synthetic-event case (added when #27937 introduced
-           ``direct_messages_topic_id`` fallback for sends without an anchor):
-           if Bot API rejects the topic id itself with any BadRequest that
-           mentions topic/thread routing, we retry without routing rather
-           than dropping the message.
-        """
-        if not (metadata and metadata.get("telegram_dm_topic_reply_fallback")):
-            return False
-        if not cls._is_bad_request_error(error):
-            return False
-        err_lower = str(error).lower()
-        if reply_to_message_id is not None and "message to be replied not found" in err_lower:
-            return True
-        # Synthetic / resumed sends route via ``direct_messages_topic_id``
-        # instead of a reply anchor. If Telegram rejects the topic id, fall
-        # back to a plain DM send.
-        if metadata.get("direct_messages_topic_id"):
-            topic_markers = (
-                "direct_messages_topic",
-                "message thread not found",
-                "thread not found",
-                "topic_closed",
-                "topic_deleted",
-                "topic not found",
-            )
-            if any(marker in err_lower for marker in topic_markers):
-                return True
-        return False
-
-    async def _send_with_dm_topic_reply_anchor_retry(
-        self,
-        send_fn: Any,
-        send_kwargs: Dict[str, Any],
-        metadata: Optional[Dict[str, Any]],
-        reply_to_message_id: Optional[int],
-        media_label: str,
-        reset_media: Optional[Any] = None,
-    ) -> Any:
-        """Retry stale private-topic media replies once without the topic anchor."""
-        try:
-            return await send_fn(**send_kwargs)
-        except Exception as send_err:
-            if not self._should_retry_without_dm_topic_reply_anchor(
-                send_err,
-                metadata,
-                reply_to_message_id,
-            ):
-                raise
-            logger.warning(
-                "[%s] Reply target deleted for Telegram %s, "
-                "retrying without reply/topic anchor: %s",
-                self.name,
-                media_label,
-                _redact_telegram_error_text(send_err),
-            )
-            if reset_media is not None:
-                reset_media()
-            retry_kwargs = dict(send_kwargs)
-            retry_kwargs["reply_to_message_id"] = None
-            retry_kwargs.pop("message_thread_id", None)
-            retry_kwargs.pop("direct_messages_topic_id", None)
-            return await send_fn(**retry_kwargs)
 
     @staticmethod
     def _looks_like_connect_timeout(error: Exception) -> bool:
@@ -1335,309 +1078,6 @@ class TelegramAdapter(TelegramMediaMixin, TelegramLifecycleMixin, TelegramReacti
         return parsed
 
 
-
-    async def _create_dm_topic(
-        self,
-        chat_id: int,
-        name: str,
-        icon_color: Optional[int] = None,
-        icon_custom_emoji_id: Optional[str] = None,
-    ) -> Optional[int]:
-        """Create a forum topic in a private (DM) chat.
-
-        Uses Bot API 9.4's createForumTopic which now works for 1-on-1 chats.
-        Returns the message_thread_id on success, None on failure.
-        """
-        if not self._bot:
-            return None
-        try:
-            kwargs: Dict[str, Any] = {"chat_id": chat_id, "name": name}
-            if icon_color is not None:
-                kwargs["icon_color"] = icon_color
-            if icon_custom_emoji_id:
-                kwargs["icon_custom_emoji_id"] = icon_custom_emoji_id
-
-            topic = await self._bot.create_forum_topic(**kwargs)
-            thread_id = topic.message_thread_id
-            logger.info(
-                "[%s] Created DM topic '%s' in chat %s -> thread_id=%s",
-                self.name, name, chat_id, thread_id,
-            )
-            return thread_id
-        except Exception as e:
-            error_text = str(e).lower()
-            # If topic already exists, try to find it via getForumTopicIconStickers
-            # or we just log and skip — Telegram doesn't provide a "list topics" API
-            if "topic_name_duplicate" in error_text or "already" in error_text:
-                logger.info(
-                    "[%s] DM topic '%s' already exists in chat %s (will be mapped from incoming messages)",
-                    self.name, name, chat_id,
-                )
-            elif "not a forum" in error_text or "forums_disabled" in error_text:
-                logger.warning(
-                    "[%s] Cannot create DM topic '%s' in chat %s: Topics mode is not enabled. "
-                    "The user must open the DM with this bot in Telegram, tap the bot name "
-                    "at the top, and enable 'Topics' in chat settings before topics can be created.",
-                    self.name, name, chat_id,
-                )
-            else:
-                logger.warning(
-                    "[%s] Failed to create DM topic '%s' in chat %s: %s",
-                    self.name, name, chat_id, _redact_telegram_error_text(e),
-                )
-            return None
-
-    async def create_handoff_thread(
-        self,
-        parent_chat_id: str,
-        name: str,
-    ) -> Optional[str]:
-        """Create a forum topic for a session handoff.
-
-        Works for DM topics (Bot API 9.4+, requires user to enable Topics
-        in their chat with the bot) and forum supergroups. Returns the
-        ``message_thread_id`` as a string, or ``None`` on failure.
-        """
-        try:
-            chat_id_int = int(parent_chat_id)
-        except (TypeError, ValueError):
-            return None
-        thread_id = await self._create_dm_topic(chat_id_int, name=name)
-        return str(thread_id) if thread_id else None
-
-    async def ensure_dm_topic(self, chat_id: str, topic_name: str, force_create: bool = False) -> Optional[str]:
-        """Return a private DM topic thread id, creating and persisting it if needed."""
-        name = str(topic_name or "").strip()
-        if not name:
-            return None
-        try:
-            chat_id_int = int(chat_id)
-        except (TypeError, ValueError):
-            return None
-
-        cache_key = f"{chat_id_int}:{name}"
-        cached = self._dm_topics.get(cache_key)
-        if cached and not force_create:
-            return str(cached)
-
-        topic_conf: Optional[Dict[str, Any]] = None
-        chat_entry: Optional[Dict[str, Any]] = None
-        for entry in self._dm_topics_config:
-            if str(entry.get("chat_id")) != str(chat_id_int):
-                continue
-            chat_entry = entry
-            for candidate in entry.get("topics", []):
-                if candidate.get("name") == name:
-                    topic_conf = candidate
-                    break
-            break
-
-        if topic_conf and topic_conf.get("thread_id") and not force_create:
-            thread_id = int(topic_conf["thread_id"])
-            self._dm_topics[cache_key] = thread_id
-            return str(thread_id)
-
-        if chat_entry is None:
-            chat_entry = {"chat_id": chat_id_int, "topics": []}
-            self._dm_topics_config.append(chat_entry)
-        if topic_conf is None:
-            topic_conf = {"name": name}
-            chat_entry.setdefault("topics", []).append(topic_conf)
-
-        thread_id = await self._create_dm_topic(
-            chat_id_int,
-            name=name,
-            icon_color=topic_conf.get("icon_color"),
-            icon_custom_emoji_id=topic_conf.get("icon_custom_emoji_id"),
-        )
-        if not thread_id:
-            return None
-
-        topic_conf["thread_id"] = thread_id
-        self._dm_topics[cache_key] = int(thread_id)
-        self._persist_dm_topic_thread_id(chat_id_int, name, int(thread_id), replace_existing=force_create)
-        return str(thread_id)
-
-    async def rename_dm_topic(
-        self,
-        chat_id: int,
-        thread_id: int,
-        name: str,
-    ) -> None:
-        """Rename a forum topic in a private (DM) chat."""
-        if not self._bot:
-            return
-        try:
-            chat_id_arg = int(chat_id)
-        except (TypeError, ValueError):
-            chat_id_arg = chat_id
-        await self._bot.edit_forum_topic(
-            chat_id=chat_id_arg,
-            message_thread_id=int(thread_id),
-            name=name,
-        )
-        logger.info(
-            "[%s] Renamed DM topic in chat %s thread_id=%s -> '%s'",
-            self.name, chat_id, thread_id, name,
-        )
-
-    def _persist_dm_topic_thread_id(
-        self,
-        chat_id: int,
-        topic_name: str,
-        thread_id: int,
-        replace_existing: bool = False,
-    ) -> None:
-        """Save a newly created thread_id back into config.yaml so it persists across restarts."""
-        try:
-            from hermes_constants import get_hermes_home
-            config_path = get_hermes_home() / "config.yaml"
-            if not config_path.exists():
-                logger.warning("[%s] Config file not found at %s, cannot persist thread_id", self.name, config_path)
-                return
-
-            import yaml as _yaml
-            with open(config_path, "r", encoding="utf-8") as f:
-                config = _yaml.safe_load(f) or {}
-
-            # Navigate to platforms.telegram.extra.dm_topics, creating the path
-            # when a named delivery target asks us to create a topic that was
-            # not predeclared in config.yaml.
-            platforms = config.setdefault("platforms", {})
-            telegram_config = platforms.setdefault("telegram", {})
-            extra = telegram_config.setdefault("extra", {})
-            dm_topics = extra.setdefault("dm_topics", [])
-
-            changed = False
-            matching_chat_entry = None
-            for chat_entry in dm_topics:
-                try:
-                    chat_matches = int(chat_entry.get("chat_id", 0)) == int(chat_id)
-                except (TypeError, ValueError):
-                    chat_matches = False
-                if not chat_matches:
-                    continue
-                matching_chat_entry = chat_entry
-                for t in chat_entry.setdefault("topics", []):
-                    if t.get("name") == topic_name:
-                        if replace_existing or not t.get("thread_id"):
-                            if t.get("thread_id") != thread_id:
-                                t["thread_id"] = thread_id
-                                changed = True
-                        break
-                else:
-                    chat_entry.setdefault("topics", []).append(
-                        {"name": topic_name, "thread_id": thread_id}
-                    )
-                    changed = True
-                break
-
-            if matching_chat_entry is None:
-                dm_topics.append({
-                    "chat_id": chat_id,
-                    "topics": [{"name": topic_name, "thread_id": thread_id}],
-                })
-                changed = True
-
-            if changed:
-                from hermes_cli.config import atomic_config_write
-
-                atomic_config_write(
-                    config_path,
-                    config,
-                    default_flow_style=False,
-                    sort_keys=False,
-                )
-                logger.info(
-                    "[%s] Persisted thread_id=%s for topic '%s' in config.yaml",
-                    self.name, thread_id, topic_name,
-                )
-        except Exception as e:
-            logger.warning("[%s] Failed to persist thread_id to config: %s", self.name, e, exc_info=True)
-
-    async def _setup_dm_topics(self) -> None:
-        """Load or create configured DM topics for specified chats.
-
-        Reads config.extra['dm_topics'] — a list of dicts:
-        [
-            {
-                "chat_id": 123456789,
-                "topics": [
-                    {"name": "General", "icon_color": 7322096, "thread_id": 100},
-                    {"name": "Accessibility Auditor", "icon_color": 9367192, "skill": "accessibility-auditor"}
-                ]
-            }
-        ]
-
-        If a topic already has a thread_id in the config (persisted from a previous
-        creation), it is loaded into the cache without calling createForumTopic.
-        Only topics without a thread_id are created via the API, and their thread_id
-        is then saved back to config.yaml for future restarts.
-        """
-        if not self._dm_topics_config:
-            return
-
-        for chat_entry in self._dm_topics_config:
-            chat_id = chat_entry.get("chat_id")
-            topics = chat_entry.get("topics", [])
-            if not chat_id or not topics:
-                continue
-
-            logger.info(
-                "[%s] Setting up %d DM topic(s) for chat %s",
-                self.name, len(topics), chat_id,
-            )
-
-            for topic_conf in topics:
-                topic_name = topic_conf.get("name")
-                if not topic_name:
-                    continue
-
-                cache_key = f"{chat_id}:{topic_name}"
-
-                # If thread_id is already persisted in config, just load into cache
-                existing_thread_id = topic_conf.get("thread_id")
-                if existing_thread_id:
-                    self._dm_topics[cache_key] = int(existing_thread_id)
-                    logger.info(
-                        "[%s] DM topic loaded from config: %s -> thread_id=%s",
-                        self.name, cache_key, existing_thread_id,
-                    )
-                    continue
-
-                # No persisted thread_id — create the topic via API
-                icon_color = topic_conf.get("icon_color")
-                icon_emoji = topic_conf.get("icon_custom_emoji_id")
-
-                thread_id = await self._create_dm_topic(
-                    chat_id=normalize_telegram_chat_id(chat_id),
-                    name=topic_name,
-                    icon_color=icon_color,
-                    icon_custom_emoji_id=icon_emoji,
-                )
-
-                if thread_id:
-                    self._dm_topics[cache_key] = thread_id
-                    logger.info(
-                        "[%s] DM topic cached: %s -> thread_id=%s",
-                        self.name, cache_key, thread_id,
-                    )
-                    # Persist thread_id to config so we don't recreate on next restart
-                    self._persist_dm_topic_thread_id(int(chat_id), topic_name, thread_id)
-
-                    # Send a seed message so the topic is visible in Telegram's client.
-                    # Empty topics are hidden by the client UI until they contain a message.
-                    try:
-                        await self._bot.send_message(
-                            chat_id=normalize_telegram_chat_id(chat_id),
-                            message_thread_id=thread_id,
-                            text=f"\U0001f4cc {topic_name}",
-                        )
-                    except Exception as seed_err:
-                        logger.debug(
-                            "[%s] Could not send seed message to topic '%s': %s",
-                            self.name, topic_name, seed_err,
-                        )
 
     async def send_update_prompt(
         self, chat_id: str, prompt: str, default: str = "",
@@ -3914,104 +3354,6 @@ class TelegramAdapter(TelegramMediaMixin, TelegramLifecycleMixin, TelegramReacti
         except Exception as exc:
             adapter_name = getattr(self, "name", "telegram")
             logger.warning("[%s] Failed to observe Telegram group message: %s", adapter_name, exc)
-
-
-    def _reload_dm_topics_from_config(self) -> None:
-        """Re-read dm_topics from config.yaml and load any new thread_ids into cache.
-
-        This allows topics created externally (e.g. by the agent via API) to be
-        recognized without a gateway restart.
-        """
-        try:
-            # Canonical loader: behavioral read (dm_topics routing) now honors
-            # managed-scope overlay + ${VAR} expansion like every other read.
-            from hermes_cli.config import load_config_readonly
-            config = load_config_readonly()
-
-            dm_topics = (
-                config.get("platforms", {})
-                .get("telegram", {})
-                .get("extra", {})
-                .get("dm_topics", [])
-            )
-            if not dm_topics:
-                # Clear both config and precomputed set when all topics are removed
-                self._dm_topics_config = []
-                self._dm_topic_chat_ids = set()
-                return
-
-            # Update in-memory config and cache any new thread_ids
-            self._dm_topics_config = dm_topics
-            # Rebuild the chat_id set for O(1) root-DM ignore lookup
-            self._dm_topic_chat_ids = {
-                str(chat_entry["chat_id"]) for chat_entry in dm_topics if "chat_id" in chat_entry
-            }
-            for chat_entry in dm_topics:
-                cid = chat_entry.get("chat_id")
-                if not cid:
-                    continue
-                for t in chat_entry.get("topics", []):
-                    tid = t.get("thread_id")
-                    name = t.get("name")
-                    if tid and name:
-                        cache_key = f"{cid}:{name}"
-                        if cache_key not in self._dm_topics:
-                            self._dm_topics[cache_key] = int(tid)
-                            logger.info(
-                                "[%s] Hot-loaded DM topic from config: %s -> thread_id=%s",
-                                self.name, cache_key, tid,
-                            )
-        except Exception as e:
-            logger.debug("[%s] Failed to reload dm_topics from config: %s", self.name, e)
-
-    def _get_dm_topic_info(self, chat_id: str, thread_id: Optional[str]) -> Optional[Dict[str, Any]]:
-        """Look up DM topic config by chat_id and thread_id.
-
-        Returns the topic config dict (name, skill, etc.) if this thread_id
-        matches a known DM topic, or None.
-        """
-        if not thread_id:
-            return None
-
-        thread_id_int = int(thread_id)
-
-        # Check cached topics first (created by us or loaded at startup)
-        for key, cached_tid in self._dm_topics.items():
-            if cached_tid == thread_id_int and key.startswith(f"{chat_id}:"):
-                topic_name = key.split(":", 1)[1]
-                # Find the full config for this topic
-                for chat_entry in self._dm_topics_config:
-                    if str(chat_entry.get("chat_id")) == chat_id:
-                        for t in chat_entry.get("topics", []):
-                            if t.get("name") == topic_name:
-                                return t
-                return {"name": topic_name}
-
-        # Not in cache — hot-reload config in case topics were added externally
-        self._reload_dm_topics_from_config()
-
-        # Check cache again after reload
-        for key, cached_tid in self._dm_topics.items():
-            if cached_tid == thread_id_int and key.startswith(f"{chat_id}:"):
-                topic_name = key.split(":", 1)[1]
-                for chat_entry in self._dm_topics_config:
-                    if str(chat_entry.get("chat_id")) == chat_id:
-                        for t in chat_entry.get("topics", []):
-                            if t.get("name") == topic_name:
-                                return t
-                return {"name": topic_name}
-
-        return None
-
-    def _cache_dm_topic_from_message(self, chat_id: str, thread_id: str, topic_name: str) -> None:
-        """Cache a thread_id -> topic_name mapping discovered from an incoming message."""
-        cache_key = f"{chat_id}:{topic_name}"
-        if cache_key not in self._dm_topics:
-            self._dm_topics[cache_key] = int(thread_id)
-            logger.info(
-                "[%s] Cached DM topic from message: %s -> thread_id=%s",
-                self.name, cache_key, thread_id,
-            )
 
 
 # ──────────────────────────────────────────────────────────────────────────
